requirements.txt
pygame==2.5.2
pillow==10.1.0
pyinstaller==6.3.0
